Don't use `update_attribute`
This method is ambiguous. Sometimes we use it to set invalid data in
tests (which can usually be done with `update_column`), and other times
we use it instead of `update!`.

I'm removing it because, even if sometimes it could make sense to use
it, it's too similar to `update_attributes` (which is an alias for
`update` and runs validations), making it confusing.

I'm not blacklisting any other method which skips validations because we
know they skip validations and use them anyway (hopefully with care).