Move performance and security rules to basic cops
As mentioned in commit 9d566a22, I've kept these performance cops but
not for performance reasons but because the make the code easier to
read.

As for the security cops, we've never had problems with any of them, but
since we recently added an `eval` call by accident, there's a chance we
could do the same with JSONLoad and YAMLLoad.