Simplify setting tsvector values
We make the code easier to read and at the same time remove a SQL
injection false positive regarding the use of `WHERE id = #{id}`.

We still get a warning about SQL injection regarding `tsv =` part. It's
a false positive, since the value of that parameter does not depend on
user input.