Use "resolve" for polymorphic hierarchy paths
Note that now the `admin_polymorphic_path` method will not work for
every model due to inconsistency in our admin routes. For instance, we
define `groups` and `budget_investments`; we should either use the
`budget_` prefix in all places or remove it everywhere. Right now the
code only works for items with the prefix, and it isn't a big deal
because we never call it with an item without the prefix.

The following code would only work for items without the prefix.

```
def admin_polymorphic_path(resource, options = {})
  resolve = resolve_for(resource)

  if resolve
    if resolve.last.is_a?(Hash)
      polymorphic_path([:admin, *resolve[0..-2]], options.merge(resolve.last))
    else
      polymorphic_path([:admin, resolve], options)
    end
  else
    polymorphic_path([:admin, resource], options)
  end
end

private

  def resolve_for(resource)
    polymorphic_mapping(resource)&.send(:eval_block, self, resource, {})
  end
```