Remove unused tag filter
This filter was added in commit 4285ba4b but most of the code from that
commit has disappeared without a trace (maybe due to a merge conflict?).

This filter could actually be useful if we started using it when users
click on a tag. Since we don't, I'm removing it. We might add it back if
we decide to actually use it.