Wait till CKEditor is ready before checking it
With chromedriver >= 80, the tests are freezing sometimes when the same
editor is loaded again.

For example, visit a page, fill in CKEditor, and then visit the same
page again and check the contents of CKEditor. The second time,
chromedriver freezes. If, on the other hand, the second time we check
the contents of a different CKEditor, everything works as expected.

We don't know whether it's a CKEditor issue or a chromedriver issue. In
the past we've had some errors related to CKEditor trying to load the
same instance twice and we aren't sure they have been fixed since we
could never reproduce them. It's also possible that chromedriver has
some sort of cache for iframes and if CKEditor isn't ready it tries to
access the frame from the previous request, which had the same locator.

Since chromedriver no longer hangs after this change, we don't have to
force any chromedriver version anymore.