Extract investments search form into a component
We had helper methods all over the place which were only used in this
view. Since we're going to change some of them to use the budget as a
parameter, they don't belong in those helpers anymore, and extracting
them to a component makes the code easier to understand and refactor.

Note the method `budget_heading_select_options` is used in more places,
so we're keeping it as a helper method for now.