Deactivate GraphQL feature by default
We added the option to deactivate it in commit 0063e7b4. We decided to
set it to "true" by default so upgrading to version 1.1 wouldn't
deactivate this feature in existing installations, making this option
backwards compatible.

However, since this is an option most people don't need, we're disabling
it by default for new installations. Existing installations using
version 1.1 will not be affected by this change when upgrading, since we
don't overwrite existing settings.