CSS sticky in legislation
We were using some JavaScript to make an element sticky, which has
caused some issues, particularly when trying to make it compatible with
Turbolinks.

Since the time we added that code (commit 287b144b), browser support for
`position: sticky` has improved considerably. Nowadays it works on 95%
of the browsers, and on the browsers where it doesn't work (like
Internet Explorer and Opera Mini), it isn't such a big issue. After all,
the effect only consists on having the table of contents visible the
whole time while scrolling through the contents of the text.