Make form_with generate ID attributes
Even if we don't use form_with, it makes sense to configure it to behave
the same way form_for does.

This is the default option in Rails 5.2 applications. IMHO it should
have been the default option for Rails 5.1 too, since generally form
inputs need an ID so they can easily be associated with a label.