Explicitly disable JS in legislation answer tests
The user experience with JavaScript enabled is actually very bad;
there's a usability issue here because it's impossible to change an
answer once a "radio button" is selected, which goes against the
standard practice on basically any HTML form.

Issue 4123 already mention this problem. Until we fix it, we're
disabling JavaScript in these tests.