Explicitly use rack driver in moderation specs
Just like in commit 5f0c422eb, the filter "Mark as viewed" doesn't work
properly, so here's a case where the test would fail with JavaScript not
because the test is wrong, but due to a bug. The test was passing
simplify because there was a typo in the CSS selector, which was
supposed to select an element by ID but didn't have the "#" prefix.

For now we're keeping the test as it was, but eventually we'll have to
fix the bug.